fix: deps cycle
