+++ packages/shared/src/graphql/comments.ts
@@ -5,7 +5,6 @@ import { EmptyResponse } from './emptyResponse';
 import { UserShortProfile } from '../lib/user';
 import type { Post, UserVote } from './posts';
 import { Company } from '../lib/userCompany';
-import { generateQueryKey, RequestKey } from '../lib/query';
 import { ContentPreference } from './contentPreference';
 
 export interface Author {
@@ -267,25 +266,4 @@ export const COMMENT_BY_ID_QUERY = gql`
 
 export type CommentFeedData = {
   page: Connection<Comment>;
-};
-
-type QueryKeyReturnType = ReturnType<typeof generateQueryKey>;
-
-interface GenerateCommentsQueryKeyProps {
-  postId: string;
-  sortBy: SortCommentsBy;
-}
-
-export const generateCommentsQueryKey = ({
-  postId,
-  sortBy,
-}: GenerateCommentsQueryKeyProps): QueryKeyReturnType =>
-  generateQueryKey(RequestKey.PostComments, null, { postId, sortBy });
-
-export const getAllCommentsQuery = (postId: string): QueryKeyReturnType[] => {
-  const sorting = Object.values(SortCommentsBy).map((sortBy) =>
-    generateCommentsQueryKey({ postId, sortBy }),
-  );
-
-  return sorting;
 };
